iaff_back/requirements.txt
Flask==2.2.2
Flask-Login==0.6.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
pika==1.3.2
pandas==2.1.3
requests==2.31.0
PyJWT==2.8.0
Werkzeug==2.2.2
python-dotenv==1.0.0